修复 parseIntArray 方法，确保正确解析整数数组

+++ java/src/main/java/org/zero123/PyScriptEngine/ModSdk/Utils.java
@@ -32,7 +32,7 @@ public class Utils
         int[] result = new int[parts.length];
         for (int i = 0; i < parts.length; i++)
         {
-            result[i] = Integer.parseInt(parts[i]);
+            result[i] = (int) Double.parseDouble(parts[i]);
         }
         return result;
     }
